requirements.txt
numpy>=1.19.0
pandas>=1.2.0
plotly>=5.0.0
wordcloud>=1.8.0
textblob>=0.15.3
urlextract>=1.0.0
emoji>=1.2.0
nltk>=3.6.0
scikit-learn>=0.24.0
python-dateutil>=2.8.0
matplotlib>=3.3.0  # Required for wordcloud 